Update spike sorting pipeline
