+++ src/nwb_datajoint/common/__init__.py
@@ -26,12 +26,10 @@ from .common_spikesorting import (SortGroup, SpikeSortingFilterParameters, Spike
                                   SpikeSortingWorkspace, 
                                   SpikeSorter, SpikeSorterParameters, SortingID,
                                   SpikeSortingSelection, SpikeSorting, 
-                                  SpikeSortingMetricParameters,
                                   ModifySortingParameters, ModifySortingSelection, ModifySorting, 
-                                  AutomaticCurationParameters, AutomaticCurationSelection,
-                                  AutomaticCuration,
-                                  CuratedSpikeSortingSelection, CuratedSpikeSorting,
-                                  UnitInclusionParameters)
+                                  )
+from .common_waveforms import (WaveformParameters, WaveformSelection, Waveforms)
+from .common_metrics import (MetricParameters, MetricSelection, QualityMetrics)
 from .common_subject import Subject
 from .common_task import Task, TaskEpoch
 from .nwb_helper_fn import (close_nwb_files, estimate_sampling_rate,
