+++ src/nwb_datajoint/common/common_ephys.py
@@ -152,6 +152,7 @@ class Raw(dj.Imported):
     -> IntervalList
     raw_object_id: varchar(80)      # the NWB object ID for loading this object from the file
     sampling_rate: float            # Sampling rate calculated from data, in Hz
+    conversion: float               # Microvolts per ADC unit
     comments: varchar(80)
     description: varchar(80)
     """
